aider: Add a function to add reviewers.
# Aider chat conversation:

USER: add a function to add reviewers

ASSISTANT: To add a function for adding reviewers to the `PeerReview` contract, we need to modify the `contracts/PeerReview.sol` file. This function will allow adding a reviewer's address and their associated keywords to the `reviewers` array. Only the owner of the contract should be able to add reviewers, similar to the restriction we placed on adding authors.

Here's how we can implement this:

1. Define a new function named `addReviewer` in the `PeerReview` contract.
2. The function should accept an address and an array of strings (keywords) as parameters.
3. Include a require statement to ensure that only the owner can call this function.
4. Push the new reviewer to the `reviewers` array.

Here is the *SEARCH/REPLACE block* for adding the `addReviewer` function:

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
    // Function to add an author, only callable by the owner
    function addAuthor(address _author) public {
        require(msg.sender == owner, "Only the owner can add authors.");
        authors.push(_author);
    }
=======
    // Function to add an author, only callable by the owner
    function addAuthor(address _author) public {
        require(msg.sender == owner, "Only the owner can add authors.");
        authors.push(_author);
    }

    // Function to add a reviewer, only callable by the owner
    function addReviewer(address _reviewer, string[] memory _keywords) public {
        require(msg.sender == owner, "Only the owner can add reviewers.");
        reviewers.push(Reviewer(_reviewer, _keywords));
    }
>>>>>>> REPLACE
```

This change introduces a new function `addReviewer` that allows the contract owner to add reviewers by specifying their address and associated keywords.