aider: Added test to add "transactions" keyword to reviewer 3 and "fees" keyword to reviewer 4.
# Aider chat conversation:

USER: write a test to add "transactions" keyword to the reviewer 3 ([2]) and "fees" keyword to the reviewer 4 ([3])

ASSISTANT: To write a test that adds the "transactions" keyword to reviewer 3 and "fees" keyword to reviewer 4, we need to modify the `test/PeerReview.t.sol` file. This involves adding a new test function that performs these actions and verifies the changes.

Here's how we can do it:

test/PeerReview.t.sol
```solidity
<<<<<<< SEARCH
}
=======
    function testAddKeywordsToReviewers() public {
        // Adding "transactions" keyword to reviewer 3
        peerReview.addKeywordToReviewer(2, "transactions");
        (, string[] memory reviewer3Keywords) = peerReview.getReviewer(2);
        assertEq(reviewer3Keywords[reviewer3Keywords.length - 1], "transactions");

        // Adding "fees" keyword to reviewer 4
        peerReview.addKeywordToReviewer(3, "fees");
        (, string[] memory reviewer4Keywords) = peerReview.getReviewer(3);
        assertEq(reviewer4Keywords[reviewer4Keywords.length - 1], "fees");
    }
>>>>>>> REPLACE
```

However, to fully implement this test, we also need to add the `addKeywordToReviewer` function to the `PeerReview` contract in `contracts/PeerReview.sol`, which allows adding a keyword to a specific reviewer. This function does not currently exist in the provided contract code. Would you like to proceed with adding this functionality to the `PeerReview` contract?